+++ mini_template/src/lib.rs
@@ -6,6 +6,7 @@ pub mod modifier;
 mod parser;
 mod renderer;
 mod template;
+mod util;
 pub mod value;
 mod value_container;